+++ trufflepig/bchain/getaccountdata.py
@@ -1,16 +1,50 @@
 import logging
-import itertools
+import multiprocessing as mp
 
 import pandas as pd
 import numpy as np
-
 from steem.account import Account
+
+import trufflepig.bchain.getdata as tpbg
+from trufflepig.utils import progressbar
 
 
 logger = logging.getLogger(__name__)
 
 
 MEMO_START = 'https://steemit.com/'
+
+BITBOTS = list({'smartmarket', 'smartsteem', 'upme', 'randowhale',
+            'minnowbooster', 'boomerang', 'booster', 'hak4life',
+            'lays', 'speedvoter', 'ebargains', 'danzy', 'bumper',
+            'upvotewhale', 'treeplanter', 'minnowpond', 'morwhale',
+            'drotto', 'postdoctor', 'moonbot', 'tipu', 'blockgators',
+            'echowhale', 'steemvote', 'byresteem', 'originalworks', 'withsmn',
+            'siditech', 'alphaprime', 'hugewhale', 'steemvoter', 'hottopic',
+            'resteemable', 'earthnation-bot', 'photocontests', 'friends-bot',
+           'followforupvotes', 'frontrunner', 'resteembot', 'steemlike',
+           'thundercurator', 'earnmoresteem', 'microbot', 'coolbot',
+           'thehumanbot', 'steemthat', 'gangvote', 'refresh', 'cabbage-dealer',
+           'growingpower', 'postresteem', 'mecurator', 'talhadogan',
+           'okankarol', 'bidseption', 'highvote', 'oguzhangazi', 'ottoman',
+           'resteemr', 'superbot', 'bestvote', 'zerotoherobot', 'red-rose',
+           'jeryalex', 'oceansbot', 'fresteem', 'otobot', 'bidbot',
+           'honestbot', 'upgoater', 'whalebuilder', 'postpromoter', 'pwrup',
+           'spydo', 'upmewhale', 'promobot', 'puppybot', 'moneymatchgaming',
+           'sneaky-ninja', 'zapzap', 'sleeplesswhale', 'estream.studios',
+           'seakraken', 'canalcrypto', 'upmyvote', 'hotbot',
+           'redlambo', 'slimwhale', 'singing.beauty', 'inciter', 'lovejuice',
+           'steembidbot', 'bid4joy', 'mitsuko', 'pushup', 'luckyvotes',
+           'discordia', 'shares', 'postdoctor', 'upboater',
+           'megabot', 'dailyupvotes', 'ebargains', 'bluebot', 'upyou',
+           'edensgarden', 'smartwhale', 'voterunner', 'nado.bot',
+           'jerrybanfield', 'foxyd', 'onlyprofitbot', 'minnowhelper',
+           'msp-bidbot', 'therising', 'bearwards', 'thebot', 'buildawhale',
+           'chronocrypto', 'brupvoter', 'smartsteem', 'payforplay',
+           'adriatik', 'cryptoempire', 'isotonic', 'minnowfairy',
+           'appreciator', 'childfund', 'mercurybot', 'allaz', 'sunrawhale',
+           'mrswhale', 'kittybot', 'lightningbolt', 'hottopic',
+           'sportic'})
 
 
 def find_nearest_index(target_datetime,
@@ -20,8 +54,6 @@ def find_nearest_index(target_datetime,
                            max_tries=5000,
                            index_tolerance=5):
     """ Finds nearest account action index to `target_datetime`
-
-    Currently NOT used in production!
 
     Parameters
     ----------
@@ -49,17 +81,17 @@ def find_nearest_index(target_datetime,
     current_index = latest_index
     best_largest_index = latest_index
 
-    action = next(acc.get_account_history(best_largest_index, limit=10))
+    action = next(acc.get_account_history(best_largest_index, limit=1))
     best_largest_datetime = pd.to_datetime(action['timestamp'])
     if target_datetime > best_largest_datetime:
-        logger.warning('Target beyond largest block num')
+        logger.debug('Target beyond largest block num')
         return latest_index, best_largest_datetime
 
     best_smallest_index = 1
     increase = index_tolerance + 1
     for _ in range(max_tries):
         try:
-            action = next(acc.get_account_history(current_index, limit=10))
+            action = next(acc.get_account_history(current_index, limit=1))
             current_datetime = pd.to_datetime(action['timestamp'])
             if increase <= index_tolerance:
                 return current_index, current_datetime
@@ -74,11 +106,15 @@ def find_nearest_index(target_datetime,
                 current_index = best_smallest_index + increase
 
                 if current_index < 0 or current_index > latest_index:
-                    raise RuntimeError('Seriously?')
+                    raise RuntimeError('Seriously? Error for '
+                                       'account {}: current_index {} '
+                                       'latest_index {}'.format(account,
+                                                                current_index,
+                                                                latest_index))
         except Exception:
             logger.exception('Problems for index {}'.format(current_index))
-            current_index -= 1
-            best_smallest_index -= 1
+            current_index += 1
+            best_largest_index += 1
 
 
 def get_delegates_and_shares(account, steem):
@@ -155,29 +191,132 @@ def get_delegate_payouts(account, steem, current_datetime,
     return payouts
 
 
-def get_upvote_payments(account, steem, min_timestamp):
+def get_upvote_payments(account, steem, min_datetime, max_datetime,
+                        batch_size=500):
 
     upvote_payments = {}
 
     acc = Account(account, steem)
-    transfers = acc.history_reverse(filter_by='transfer')
+    start_index, _ = find_nearest_index(max_datetime,
+                                     account, steem)
+    try:
+        transfers = history_reverse(account, steem, filter_by='transfer',
+                                    start_index=start_index,
+                                    batch_size=batch_size)
+    except Exception as e:
+        logger.exception('Could not get account data from {}'.format(account))
+        transfers = []
 
     for transfer in transfers:
         try:
             memo = transfer['memo']
             if memo.startswith(MEMO_START):
                 author, permalink = memo.split('/')[-2:]
-                assert author.startswith('@')
-                author = author[1:]
-                if (author, permalink) not in upvote_payments:
-                    upvote_payments[(author, permalink)] = []
-                upvote_payments[(author, permalink)].append(transfer['amount'])
+                if author.startswith('@'):
+                    author = author[1:]
+                    if (author, permalink) not in upvote_payments:
+                        upvote_payments[(author, permalink)] = []
+                    upvote_payments[(author, permalink)].append(transfer['amount'])
 
             timestamp = pd.to_datetime(transfer['timestamp'])
-            if timestamp < min_timestamp:
+            if timestamp < min_datetime:
                 break
 
         except Exception as e:
             logger.exception('Could not parse {}'.format(transfer))
 
     return upvote_payments
+
+
+def history_reverse(account, steem, start_index, filter_by=None,
+                    batch_size=10000, raw_output=False):
+        """ Stream account history in reverse chronological order."""
+        acc = Account(account, steem)
+        i = start_index
+        if batch_size > start_index:
+            batch_size = start_index
+        while i > 0:
+            if i - batch_size < 0:
+                batch_size = i
+            yield from acc.get_account_history(
+                index=i,
+                limit=batch_size,
+                order=-1,
+                filter_by=filter_by,
+                raw_output=raw_output,
+            )
+            i -= (batch_size + 1)
+
+
+def extend_upvotes_and_payments(upvote_payments, new_payments):
+    for author_permalink, upvote_list in new_payments.items():
+            if author_permalink not in upvote_payments:
+                upvote_payments[author_permalink] = []
+            upvote_payments[author_permalink].extend(upvote_list)
+    return upvote_payments
+
+
+def _get_upvote_payments_parrallel(accounts, steem_args, min_datetime,
+                                   max_datetime):
+    steem = tpbg.check_and_convert_steem(steem_args)
+    results = {}
+    for account in accounts:
+        result = get_upvote_payments(account, steem, min_datetime, max_datetime)
+        results = extend_upvotes_and_payments(results, result)
+
+    return result
+
+
+def get_upvote_payments_for_accounts(accounts, steem_args, min_datetime, max_datetime,
+                                     chunksize=10, ncores=20, timeout=1200):
+    logger.info('Querying upvote purchases between {} and '
+                '{} for {} accounts'.format(min_datetime,
+                                            max_datetime,
+                                            len(accounts)))
+    if ncores > 1:
+        chunks = [accounts[irun: irun + chunksize]
+                  for irun in range(0, len(accounts), chunksize)]
+
+        ctx = mp.get_context('spawn')
+        pool = ctx.Pool(ncores, initializer=tpbg.config_mp_logging)
+
+        async_results = []
+        for idx, chunk in enumerate(chunks):
+            result = pool.apply_async(_get_upvote_payments_parrallel,
+                                      args=(chunk, steem_args,
+                                            min_datetime, max_datetime))
+            async_results.append(result)
+
+        pool.close()
+
+        upvote_payments = {}
+        for kdx, async in enumerate(async_results):
+            try:
+                payments = async.get(timeout=timeout)
+                upvote_payments = extend_upvotes_and_payments(upvote_payments,
+                                                              payments)
+                if progressbar(kdx, len(chunks), percentage_step=5, logger=logger):
+                    logger.info('Finished chunk {} '
+                                'out of {} found so far {} '
+                                'upvote buyers...'.format(kdx + 1, len(chunks), len(upvote_payments)))
+            except TimeoutError:
+                logger.exception('Something went totally wrong dude!')
+
+        pool.join()
+    else:
+        return _get_upvote_payments_parrallel(accounts, steem_args, min_datetime,
+                                              max_datetime)
+
+    logger.info('Found {} upvote bought articles'.format(len(upvote_payments)))
+    return upvote_payments
+
+
+def get_upvote_payments_to_bots(steem_args, min_datetime, max_datetime,
+                                bots=BITBOTS, ncores=64):
+    logger.info('Getting payments to following bots {}'.format(bots))
+    return get_upvote_payments_for_accounts(accounts=bots,
+                                            steem_args=steem_args,
+                                            min_datetime=min_datetime,
+                                            max_datetime=max_datetime,
+                                            ncores=ncores,
+                                            chunksize=1)